Adding support for Go modules
Rename module to go-bot-api.
Refactor the unit tests to use white box testing style instead of
black box testing style by renaming the package name to be
the same as the main package. This avoids having the new
Go module dependency attempt to download a wrong matching
version of the library. The unit tests are tightly bound to this
version of the library.